+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/Versions.kt
@@ -17,6 +17,7 @@ internal object GradleVersions {
     val CURRENT: SemVer = SemVer.parse(GradleVersion.current().version)
     val VERSION_7_4: SemVer = SemVer.parse("7.4")
     val VERSION_7_5: SemVer = SemVer.parse("7.5")
+    val VERSION_8_0: SemVer = SemVer.parse("8.0")
 }
 
 internal object SentryVersions {