+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginSourceContextNonAndroidTest.kt
@@ -1,5 +1,6 @@
 package io.sentry.android.gradle.integration
 
+import io.sentry.android.gradle.util.GradleVersions
 import io.sentry.android.gradle.verifySourceBundleContents
 import io.sentry.android.gradle.withDummyCustomFile
 import io.sentry.android.gradle.withDummyJavaFile
@@ -8,6 +9,8 @@ import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import org.gradle.testkit.runner.TaskOutcome.SKIPPED
 import org.gradle.util.GradleVersion
+import org.hamcrest.CoreMatchers.`is`
+import org.junit.Assume.assumeThat
 import org.junit.Test
 
 class SentryPluginSourceContextNonAndroidTest :
@@ -132,6 +135,54 @@ class SentryPluginSourceContextNonAndroidTest :
             testProjectDir.root,
             "files/_/_/io/other/TestCustom.jvm",
             customContents,
+            variant = "java",
+            archivePath = "app/build/libs/app.jar"
+        )
+    }
+
+    @Test
+    fun `respects configuration cache`() {
+        assumeThat(
+            "SentryExternalDependenciesReportTask only supports " +
+                "configuration cache from Gradle 7.5 onwards",
+            GradleVersions.CURRENT >= GradleVersions.VERSION_7_5,
+            `is`(true)
+        )
+        appBuildFile.writeText(
+            // language=Groovy
+            """
+            plugins {
+              id "java"
+              id "io.sentry.jvm.gradle"
+            }
+
+            sentry {
+              debug = true
+              includeSourceContext = true
+              autoUploadSourceContext = false
+              autoUploadProguardMapping = false
+              org = "sentry-sdks"
+              projectName = "sentry-android"
+            }
+            """.trimIndent()
+        )
+
+        sentryPropertiesFile.writeText("")
+
+        val javaContents = testProjectDir.withDummyJavaFile()
+
+        val result = runner
+            .appendArguments("app:assemble")
+            .appendArguments("--configuration-cache")
+            .build()
+
+        assertTrue { "Configuration cache entry stored." in result.output }
+        assertTrue { "BUILD SUCCESSFUL" in result.output }
+
+        verifySourceBundleContents(
+            testProjectDir.root,
+            "files/_/_/com/example/TestJava.jvm",
+            javaContents,
             variant = "java",
             archivePath = "app/build/libs/app.jar"
         )